libstore: Make Store::querySubstitutablePathInfos async
(cherry picked from commit d2798d78bffd3484e990d713a7cfd31955ae4918)
